Extract acp_connection
